DB User/ Register ACtivity Update (mit float)

+++ app/src/main/java/com/example/testapp/DB_user.java
@@ -36,7 +36,7 @@ public class DB_user extends SQLiteOpenHelper {
     // TODO: Do we need user_id?
     // insert Data into Database, return true if successful
     // TODO: balance dem konstruktor zufügen
-    public Boolean insertData (String username, String password, int balance){
+    public Boolean insertData (String username, String password, float balance){
         SQLiteDatabase db = this.getWritableDatabase();
         ContentValues contentValues = new ContentValues();
         contentValues.put("user_name", username);
